+++ km3pipe/dataclasses.pyx
@@ -32,9 +32,9 @@ __maintainer__ = "Tamas Gal and Moritz Lotze"
 __email__ = "tgal@km3net.de"
 __status__ = "Development"
 __all__ = ('EventInfo', 'Point', 'Position', 'Direction', 'HitSeries',
-           'TimesliceHitSeries', 'Hit',
-           'Track', 'TrackSeries', 'Serialisable', 'SummaryframeInfo',
-           'BinaryStruct')
+           'RawHitSeries', 'TimesliceHitSeries', 'Hit', 'McHit', 'McTrack',
+           'McTrackSeries', 'Track', 'TrackSeries', 'Serialisable',
+           'SummaryframeInfo', 'BinaryStruct')
 
 
 IS_CC = {
@@ -107,7 +107,7 @@ class Convertible(object):
 class SummarysliceInfo(with_metaclass(Serialisable)):
     """JDAQSummaryslice Metadata.
     """
-    h5loc = '/'
+    h5loc = '/summary_slice'
     dtype = np.dtype([
         ('det_id', '<i4'),
         ('frame_index', '<u4'),
@@ -168,7 +168,7 @@ class SummarysliceInfo(with_metaclass(Serialisable)):
 class TimesliceInfo(with_metaclass(Serialisable)):
     """JDAQTimeslice metadata.
     """
-    h5loc = '/'
+    h5loc = '/time_slice_info'
     dtype = np.dtype([
         ('dom_id', '<u4'),
         ('frame_id', '<u4'),
@@ -228,7 +228,7 @@ class TimesliceInfo(with_metaclass(Serialisable)):
 class TimesliceFrameInfo(with_metaclass(Serialisable)):
     """JDAQTimeslice frame metadata.
     """
-    h5loc = '/'
+    h5loc = '/time_slice_frame_info'
     dtype = np.dtype([
         ('dom_id', '<u4'),
         ('fifo_status', '<u4'),
@@ -301,7 +301,7 @@ class TimesliceFrameInfo(with_metaclass(Serialisable)):
 class SummaryframeInfo(with_metaclass(Serialisable)):
     """JDAQSummaryslice frame metadata.
     """
-    h5loc = '/'
+    h5loc = '/summary_slice_info'
     dtype = np.dtype([
         ('dom_id', '<u4'),
         ('fifo_status', '<u4'),
@@ -373,6 +373,7 @@ class SummaryframeInfo(with_metaclass(Serialisable)):
 class EventInfo(object):
     """Event Metadata.
     """
+    h5loc = '/event_info'
     dtype = np.dtype([
         ('det_id', '<i4'),
         ('frame_index', '<u4'),
@@ -382,7 +383,6 @@ class EventInfo(object):
         ('n_events_gen', '<u8'),
         ('n_files_gen', '<u8'),
         ('overlays', '<u4'),
-        # ('run_id', '<u4'),
         ('trigger_counter', '<u8'),
         ('trigger_mask', '<u8'),
         ('utc_nanoseconds', '<u8'),
@@ -390,6 +390,7 @@ class EventInfo(object):
         ('weight_w1', '<f8'),
         ('weight_w2', '<f8'),
         ('weight_w3', '<f8'),
+        ('run_id', '<u4'),
         ('event_id', '<u4'),
     ])
 
@@ -429,6 +430,7 @@ class EventInfo(object):
 
     def __str__(self):
         return "Event #{0}:\n" \
+               "    run id:          {10}\n" \
                "    detector id:     {1}\n" \
                "    frame index:     {2}\n" \
                "    UTC seconds:     {3}\n" \
@@ -442,7 +444,7 @@ class EventInfo(object):
                        self.frame_index, self.utc_seconds,
                        self.utc_nanoseconds, self.mc_id, self.mc_t,
                        self.overlays, self.trigger_counter, self.trigger_mask,
-                       # self.run_id,
+                       self.run_id,
                        )
 
     def __repr__(self):
@@ -649,7 +651,7 @@ cdef class TimesliceHit:
         return self.__str__()
 
 
-cdef class MCHit:
+cdef class McHit:
     """Monte Carlo Hit on a PMT.
 
     Parameters
@@ -688,7 +690,7 @@ cdef class MCHit:
         self.triggered = triggered
 
     def __str__(self):
-        return "Hit: channel_id({0}), dom_id({1}), pmt_id({2}), tot({3}), " \
+        return "Mc Hit: channel_id({0}), dom_id({1}), pmt_id({2}), tot({3}), " \
                "time({4}), triggered({5})" \
                .format(self.channel_id, self.dom_id, self.pmt_id, self.tot,
                        self.time, self.triggered)
@@ -759,6 +761,7 @@ cdef class Track:
 
 class RawHitSeries(object):
     """A collection of hits without any calibration parameter."""
+    h5loc = '/raw_hits'
     dtype = np.dtype([
         ('channel_id', 'u1'),
         ('dom_id', '<u4'),
@@ -909,9 +912,69 @@ class RawHitSeries(object):
         return '\n'.join([str(hit) for hit in self._hits])
 
 
-class HitSeries(object):
+cdef class McTrack:
+    """Monte Carlo Particle track.
+
+    Parameters
+    ----------
+    bjorkeny : float
+    dir : Direction or numpy.ndarray
+    energy : float
+    id : int
+    interaction_channel : int
+    is_cc : int
+    length : float
+    pos : Position or numpy.ndarray
+    time : int
+    type : int
+    """
+    cdef public int id, time, type, interaction_channel
+    cdef public float energy, length, bjorkeny
+    cdef public unsigned short int is_cc
+    cdef public np.ndarray pos
+    cdef public np.ndarray dir
+
+    def __cinit__(self,
+                  float bjorkeny,
+                  dir,
+                  float energy,
+                  int id,
+                  np.int64_t interaction_channel,
+                  unsigned short int is_cc,
+                  float length,
+                  pos,
+                  int time,
+                  int type,
+                  int event_id,
+                  ):
+        self.bjorkeny = bjorkeny
+        self.is_cc = is_cc
+        self.dir = dir
+        self.energy = energy
+        self.interaction_channel = interaction_channel
+        self.id = id
+        self.length = length
+        self.pos = pos
+        self.time = time
+        self.type = type
+        self.event_id = event_id
+
+    def __str__(self):
+        return "Track: pos({0}), dir({1}), t={2}, E={3}, type={4} ({5})" \
+               .format(self.pos, self.dir, self.time, self.energy,
+                       self.type, pdg2name(self.type))
+
+    def __repr__(self):
+        return self.__str__()
+
+    def __insp__(self):
+        return self.__str__()
+
+
+class McHitSeries(object):
     """Collection of multiple Hits.
     """
+    h5loc = '/mc_hits'
     dtype = np.dtype([
         ('channel_id', 'u1'),
         ('dir_x', '<f8'),
@@ -923,7 +986,305 @@ class HitSeries(object):
         ('pos_x', '<f8'),
         ('pos_y', '<f8'),
         ('pos_z', '<f8'),
-        # ('run_id', '<u4'),
+        ('t0', '<i4'),
+        ('time', '<i4'),
+        ('tot', 'u1'),
+        ('triggered', 'u1'),
+        ('event_id', '<u4'),
+    ])
+
+    def __init__(self, arr, h5loc='/'):
+        self._arr = arr
+        self._index = 0
+        self._hits = None
+        self.h5loc = h5loc
+
+    @classmethod
+    def from_aanet(cls, hits, event_id):
+        try:
+            return cls(np.array([(
+                h.channel_id,
+                np.nan,     # h.dir.x,
+                np.nan,     # h.dir.y,
+                np.nan,     # h.dir.z,
+                h.dom_id,
+                h.id,
+                h.pmt_id,
+                np.nan,     # h.pos.x,
+                np.nan,     # h.pos.y,
+                np.nan,     # h.pos.z,
+                0,          # t0
+                h.t,
+                h.tot,
+                h.trig,
+                event_id,
+            ) for h in hits], dtype=cls.dtype))
+        except ValueError:
+            # Older aanet format.
+            return cls(np.array([(
+                ord(h.channel_id),
+                np.nan,     # h.dir.x,
+                np.nan,     # h.dir.y,
+                np.nan,     # h.dir.z,
+                h.dom_id,
+                h.id,
+                h.pmt_id,
+                np.nan,     # h.pos.x,
+                np.nan,     # h.pos.y,
+                np.nan,     # h.pos.z,
+                0,          # t0
+                h.t,
+                h.tot,
+                h.trig,
+                event_id,
+            ) for h in hits], dtype=cls.dtype))
+
+    @classmethod
+    def from_evt(cls, hits, event_id):
+        return cls(np.array([(
+            0,     # channel_id
+            np.nan,
+            np.nan,
+            np.nan,
+            0,     # dom_id
+            h.id,
+            h.pmt_id,
+            np.nan,
+            np.nan,
+            np.nan,
+            0,      # t0
+            h.time,
+            h.tot,
+            0,     # triggered
+            event_id,
+        ) for h in hits], dtype=cls.dtype))
+
+    @classmethod
+    def from_arrays(cls, channel_ids, dir_xs, dir_ys, dir_zs, dom_ids, ids,
+                    pmt_ids, pos_xs, pos_ys, pos_zs, t0s, times, tots,
+                    triggereds, event_id):
+        # do we need shape[0] or does len() work too?
+        try:
+            length = channel_ids.shape[0]
+        except AttributeError:
+            length = len(channel_ids)
+        hits = np.empty(length, cls.dtype)
+        hits['channel_id'] = channel_ids
+        hits['dir_x'] = dir_xs
+        hits['dir_y'] = dir_ys
+        hits['dir_z'] = dir_zs
+        hits['dom_id'] = dom_ids
+        hits['id'] = ids
+        hits['pmt_id'] = pmt_ids
+        hits['pos_x'] = pos_xs
+        hits['pos_y'] = pos_ys
+        hits['pos_z'] = pos_zs
+        hits['t0'] = t0s
+        hits['time'] = times
+        hits['tot'] = tots
+        hits['triggered'] = triggereds
+        hits['event_id'] = np.full(length, event_id, dtype='<u4')
+        return cls(hits)
+
+    @classmethod
+    def from_dict(cls, map, event_id):
+        if event_id is None:
+            event_id = map['event_id']
+        return cls.from_arrays(
+            map['channel_id'],
+            map['dir_x'],
+            map['dir_y'],
+            map['dir_z'],
+            map['dom_id'],
+            map['id'],
+            map['pmt_id'],
+            map['pos_x'],
+            map['pos_y'],
+            map['pos_z'],
+            map['t0'],
+            map['time'],
+            map['tot'],
+            map['triggered'],
+            event_id,
+        )
+
+    @classmethod
+    def from_table(cls, table, event_id):
+        if event_id is None:
+            event_id = table[0]['event_id']
+        return cls(np.array([(
+            row['channel_id'],
+            row['dir_x'],
+            row['dir_y'],
+            row['dir_z'],
+            row['dom_id'],
+            row['id'],
+            row['pmt_id'],
+            row['pos_x'],
+            row['pos_y'],
+            row['pos_z'],
+            row['t0'],
+            row['time'],
+            row['tot'],
+            row['triggered'],
+        ) for row in table], dtype=cls.dtype), event_id)
+
+    @classmethod
+    def deserialise(cls, *args, **kwargs):
+        return cls.conv_from(*args, **kwargs)
+
+    def serialise(self, *args, **kwargs):
+        return self.conv_to(*args, **kwargs)
+
+    @classmethod
+    def conv_from(cls, data, event_id=None, fmt='numpy', h5loc='/'):
+        # what is event_id doing here?
+        if fmt == 'numpy':
+            return cls(data)
+        if fmt == 'pandas':
+            return cls(data.to_records(index=False))
+
+    def conv_to(self, to='numpy'):
+        if to == 'numpy':
+            return KM3Array(np.array(self.__array__(), dtype=self.dtype),
+                            h5loc=self.h5loc)
+        if to == 'pandas':
+            return KM3DataFrame(self.conv_to(to='numpy'), h5loc=self.h5loc)
+
+    def __array__(self):
+        return self._arr
+
+    def __iter__(self):
+        return self
+
+    @property
+    def id(self):
+        return self._arr['id']
+
+    @property
+    def time(self):
+        return self._arr['time']
+
+    @property
+    def triggered_hits(self):
+        return McHitSeries(self._arr[self._arr['triggered'] == True])     # noqa
+
+    @property
+    def first_hits(self):
+        h = self.conv_to(to='pandas')
+        h.sort_values('time', inplace=True)
+        h = h.drop_duplicates(subset='dom_id')
+        return McHitSeries(h.to_records(index=False))
+
+    @property
+    def triggered(self):
+        return self._arr['triggered']
+
+    @property
+    def tot(self):
+        return self._arr['tot']
+
+    @property
+    def dom_id(self):
+        return self._arr['dom_id']
+
+    @property
+    def pmt_id(self):
+        return self._arr['pmt_id']
+
+    @property
+    def id(self):
+        return self._arr['id']
+
+    @property
+    def channel_id(self):
+        return self._arr['channel_id']
+
+    @property
+    def pos_x(self):
+        return self._arr['pos_x']
+
+    @property
+    def pos_y(self):
+        return self._arr['pos_y']
+
+    @property
+    def pos_z(self):
+        return self._arr['pos_z']
+
+    @property
+    def dir_x(self):
+        return self._arr['dir_x']
+
+    @property
+    def dir_y(self):
+        return self._arr['dir_y']
+
+    @property
+    def dir_z(self):
+        return self._arr['dir_z']
+
+    @property
+    def t0(self):
+        return self._arr['t0']
+
+    def next(self):
+        """Python 2/3 compatibility for iterators"""
+        return self.__next__()
+
+    def __next__(self):
+        if self._index >= len(self):
+            self._index = 0
+            raise StopIteration
+        hit = self[self._index]
+        self._index += 1
+        return hit
+
+    def __len__(self):
+        return self._arr.shape[0]
+
+    def __getitem__(self, index):
+        if isinstance(index, int):
+            hit = McHit(*self._arr[index])
+            return hit
+        elif isinstance(index, slice):
+            return self._slice_generator(index)
+        else:
+            raise TypeError("index must be int or slice")
+
+    def _slice_generator(self, index):
+        """A simple slice generator for iterations"""
+        start, stop, step = index.indices(len(self))
+        for i in range(start, stop, step):
+            yield McHit(*self._arr[i])
+
+    def __str__(self):
+        n_hits = len(self)
+        plural = 's' if n_hits > 1 or n_hits == 0 else ''
+        return("McHitSeries with {0} hit{1}.".format(len(self), plural))
+
+    def __repr__(self):
+        return self.__str__()
+
+    def __insp__(self):
+        return '\n'.join([str(hit) for hit in self._hits])
+
+
+class HitSeries(object):
+    """Collection of multiple Hits.
+    """
+    h5loc = '/hits'
+    dtype = np.dtype([
+        ('channel_id', 'u1'),
+        ('dir_x', '<f8'),
+        ('dir_y', '<f8'),
+        ('dir_z', '<f8'),
+        ('dom_id', '<u4'),
+        ('id', '<u4'),
+        ('pmt_id', '<u4'),
+        ('pos_x', '<f8'),
+        ('pos_y', '<f8'),
+        ('pos_z', '<f8'),
         ('t0', '<i4'),
         ('time', '<i4'),
         ('tot', 'u1'),
@@ -1211,6 +1572,7 @@ class HitSeries(object):
 class TimesliceHitSeries(object):
     """Collection of multiple timeslice hits.
     """
+    h5loc = '/time_slice_hits'
     dtype = np.dtype([
         ('channel_id', 'u1'),
         ('dom_id', '<u4'),
@@ -1346,13 +1708,14 @@ class TimesliceHitSeries(object):
         return '\n'.join([str(hit) for hit in self._hits])
 
 
-class TrackSeries(object):
-    """Collection of multiple Tracks.
+class McTrackSeries(object):
+    """Collection of multiple McTracks.
 
     Attributes
     ----------
     dtype: datatype of array representation
     """
+    h5loc = '/tracks'
     dtype = np.dtype([
         ('bjorkeny', '<f8'),
         ('dir_x', '<f8'),
@@ -1366,7 +1729,310 @@ class TrackSeries(object):
         ('pos_x', '<f8'),
         ('pos_y', '<f8'),
         ('pos_z', '<f8'),
-        # ('run_id', '<u4'),
+        ('time', '<i4'),
+        ('type', '<i4'),
+        ('event_id', '<u4'),
+    ])
+
+    def __init__(self, tracks, event_id, h5loc='/'):
+        self._bjorkeny = None
+        self._dir = None
+        self._energy = None
+        self._highest_energetic_muon = None
+        self._id = None
+        self._index = 0
+        self._interaction_channel = None
+        self._is_cc = None
+        self._length = None
+        self._pos = None
+        self._time = None
+        self._tracks = tracks
+        self._type = None
+        self.event_id = event_id
+        self.h5loc = h5loc
+
+    @classmethod
+    def from_aanet(cls, tracks, event_id):
+        return cls([McTrack(cls.get_usr_name(t, str('by'), 1),               # bjorkeny
+                          Direction((t.dir.x, t.dir.y, t.dir.z)),
+                          t.E,
+                          t.id,
+                          cls.get_usr_name(t, str('ichan'), 2),               # ichan
+                          IS_CC[cls.get_usr_name(t, str('cc'), 0)],        # is_cc
+                          cls.get_len(t),
+                          Position((t.pos.x, t.pos.y, t.pos.z)),
+                          t.t,
+                          # This is a nasty bug. It is not completely clear
+                          # if this is supposed to be PDG or Geant convention.
+                          # might be, that for CC neutrino events,
+                          # the two vector elements might follow _different_
+                          # conventions. Yep, 2 conventions for
+                          # 2 vector elements...
+                          #
+                          # UPDATE 2017-03/21: aanet now has all casted to PDG
+                          #
+                          # geant2pdg(t.type))
+                          t.type,
+                          )
+                    for t in tracks], event_id)
+
+    @classmethod
+    def from_arrays(cls,
+                    bjorkenys,
+                    directions_x,
+                    directions_y,
+                    directions_z,
+                    energies,
+                    ids,
+                    interaction_channels,
+                    is_ccs,
+                    lengths,
+                    positions_x,
+                    positions_y,
+                    positions_z,
+                    times,
+                    types,
+                    event_id,
+                    ):
+        directions = np.column_stack((directions_x, directions_y,
+                                      directions_z))
+        positions = np.column_stack((positions_x, positions_y,
+                                     positions_z))
+        args = bjorkenys, directions, energies, \
+            ids, interaction_channels, is_ccs, lengths, positions, \
+            times, types
+        tracks = cls([McTrack(*track_args) for track_args in zip(*args)],
+                     event_id)
+        tracks._bjorkeny = bjorkenys
+        tracks._dir = zip(directions_x, directions_y, directions_z)
+        tracks._energy = energies
+        tracks._id = ids
+        tracks._interaction_channel = interaction_channels
+        tracks._is_cc = is_ccs
+        tracks._length = lengths
+        tracks._pos = zip(positions_x, positions_y, positions_z)
+        tracks._time = times
+        tracks._type = types
+        return tracks
+
+    @classmethod
+    def from_km3df(cls, km3df):
+        return cls.from_table(km3df.conv_to(to='numpy'),
+                              event_id=int(km3df['event_id']))
+
+    @classmethod
+    def from_table(cls, table, event_id):
+        return cls([McTrack(
+            row['bjorkeny'],
+            np.array((row['dir_x'], row['dir_y'], row['dir_z'],)),
+            row['energy'],
+            row['id'],
+            row['interaction_channel'],
+            row['is_cc'],
+            row['length'],
+            np.array((row['pos_x'], row['pos_y'], row['pos_z'],)),
+            row['time'],
+            row['type']
+        ) for row in table], event_id)
+
+    @classmethod
+    def deserialise(cls, *args, **kwargs):
+        return cls.conv_from(*args, **kwargs)
+
+    def serialise(self, *args, **kwargs):
+        return self.conv_to(*args, **kwargs)
+
+    @classmethod
+    def conv_from(cls, data, event_id, fmt='numpy', h5loc='/'):
+        if fmt == 'numpy':
+            return cls.from_table(data, event_id)
+
+    def conv_to(self, to='numpy'):
+        if to == 'numpy':
+            return KM3Array(np.array(self.__array__(), dtype=self.dtype),
+                            h5loc=self.h5loc)
+        if to == 'pandas':
+            return KM3DataFrame(self.conv_to(to='numpy'), h5loc=self.h5loc)
+
+    def __array__(self):
+        return [(
+            t.bjorkeny, t.dir[0], t.dir[1], t.dir[2], t.energy,
+            t.id, t.interaction_channel, t.is_cc,
+            t.length, t.pos[0], t.pos[1], t.pos[2], t.time, t.type,
+            self.event_id,
+        ) for t in self._tracks]
+
+    @classmethod
+    def get_len(cls, track):
+        try:
+            return track.len
+        except AttributeError:
+            return 0
+
+    @classmethod
+    def get_usr_item(cls, track, index=None):
+        try:
+            item = track.usr[index]
+        except IndexError:
+            item = 0.
+        return item
+
+    @classmethod
+    def get_usr_name(cls, track, name, index=None):
+        """Try to retrieve item based on name from aanet.
+
+        If that fails (old aanet), try getting it via index.
+        """
+        try:
+          name_len = len(track.usr_names)
+        except AttributeError:
+          name_len = 0
+        if len(track.usr) > name_len:
+            return cls.get_usr_item(track, index)
+        try:
+          out = track.getusr(name)
+        except (AttributeError, KeyError):
+          out = 0
+        return out
+
+    @property
+    def highest_energetic_muon(self):
+        if self._highest_energetic_muon is None:
+            muons = [track for track in self if abs(track.type) == 13]
+            if len(muons) == 0:
+                raise AttributeError("No muon found")
+            self._highest_energetic_muon = max(muons, key=lambda m: m.energy)
+        return self._highest_energetic_muon
+
+    def __iter__(self):
+        return self
+
+    def __iter__(self):
+        return self
+
+    @property
+    def bjorkeny(self):
+        if self._bjorkeny is None:
+            self._bjorkeny = np.array([t.bjorkeny for t in self._tracks])
+        return self._bjorkeny
+
+    @property
+    def is_cc(self):
+        if self._is_cc is None:
+            self._is_cc = np.array([t.is_cc for t in self._tracks])
+        return self._is_cc
+
+    @property
+    def interaction_channel(self):
+        if self._interaction_channel is None:
+            self._interaction_channel = np.array([t.interaction_channel for
+                                                  t in self._tracks])
+        return self._interaction_channel
+
+    @property
+    def id(self):
+        if self._id is None:
+            self._id = np.array([t.id for t in self._tracks])
+        return self._id
+
+    @property
+    def time(self):
+        if self._time is None:
+            self._time = np.array([t.time for t in self._tracks])
+        return self._time
+
+    @property
+    def energy(self):
+        if self._energy is None:
+            self._energy = np.array([t.energy for t in self._tracks])
+        return self._energy
+
+    @property
+    def length(self):
+        if self._length is None:
+            self._length = np.array([t.length for t in self._tracks])
+        return self._length
+
+    @property
+    def type(self):
+        if self._type is None:
+            self._type = np.array([t.type for t in self._tracks])
+        return self._type
+
+    @property
+    def pos(self):
+        if self._pos is None:
+            self._pos = np.array([t.pos for t in self._tracks])
+        return self._pos
+
+    @property
+    def dir(self):
+        if self._dir is None:
+            self._dir = np.array([t.dir for t in self._tracks])
+        return self._dir
+
+    def next(self):
+        """Python 2/3 compatibility for iterators"""
+        return self.__next__()
+
+    def __next__(self):
+        if self._index >= len(self):
+            self._index = 0
+            raise StopIteration
+        item = self._tracks[self._index]
+        self._index += 1
+        return item
+
+    def __len__(self):
+        return len(self._tracks)
+
+    def __getitem__(self, index):
+        if isinstance(index, int):
+            return self._tracks[index]
+        elif isinstance(index, slice):
+            return self._slice_generator(index)
+        else:
+            raise TypeError("index must be int or slice")
+
+    def _slice_generator(self, index):
+        """A simple slice generator for iterations"""
+        start, stop, step = index.indices(len(self))
+        for i in range(start, stop, step):
+            yield self._tracks[i]
+
+    def __str__(self):
+        n_tracks = len(self)
+        plural = 's' if n_tracks > 1 or n_tracks == 0 else ''
+        return("McTrackSeries with {0} track{1}.".format(len(self), plural))
+
+    def __repr__(self):
+        return self.__str__()
+
+    def __insp__(self):
+        return '\n'.join([str(track) for track in self._tracks])
+
+
+class TrackSeries(object):
+    """Collection of multiple Tracks.
+
+    Attributes
+    ----------
+    dtype: datatype of array representation
+    """
+    h5loc = '/tracks'
+    dtype = np.dtype([
+        ('bjorkeny', '<f8'),
+        ('dir_x', '<f8'),
+        ('dir_y', '<f8'),
+        ('dir_z', '<f8'),
+        ('energy', '<f8'),
+        ('id', '<u4'),
+        ('interaction_channel', '<u4'),
+        ('is_cc', '<u4'),
+        ('length', '<f8'),
+        ('pos_x', '<f8'),
+        ('pos_y', '<f8'),
+        ('pos_z', '<f8'),
         ('time', '<i4'),
         ('type', '<i4'),
         ('event_id', '<u4'),
@@ -1653,6 +2319,7 @@ class TrackSeries(object):
 class SummaryframeSeries(object):
     """Collection of summary frames.
     """
+    h5loc = '/summary_frame_series'
     dtype = np.dtype([
         ('dom_id', '<u4'),
         ('max_sequence_number', '<u4'),
@@ -1870,8 +2537,9 @@ class KM3DataFrame(pd.DataFrame):
 
 
 deserialise_map = {
-    'McHits': HitSeries,
-    'Hits': RawHitSeries,
+    'McHits': McHitSeries,
+    'RawHits': RawHitSeries,
+    'Hits': HitSeries,
     'TimesliceHits': TimesliceHitSeries,
     'McTracks': TrackSeries,
     'EventInfo': EventInfo,