+++ km3pipe/tests/test_dataclasses.py
@@ -582,7 +582,9 @@ class TestSummarysliceInfo(TestCase):
 
 class TestEventInfo(TestCase):
     def test_event_info(self):
-        e = EventInfo(tuple(range(16)))
+        e = EventInfo(tuple(range(17)))
+        print(e.trigger_counter)
+        print(e)
         self.assertAlmostEqual(0, e.det_id)
         self.assertAlmostEqual(1, e.frame_index)
         self.assertAlmostEqual(2, e.livetime_sec)
@@ -591,7 +593,6 @@ class TestEventInfo(TestCase):
         self.assertAlmostEqual(5, e.n_events_gen)
         self.assertAlmostEqual(6, e.n_files_gen)
         self.assertAlmostEqual(7, e.overlays)
-        # self.assertAlmostEqual(6, e.run_id)
         self.assertAlmostEqual(8, e.trigger_counter)
         self.assertAlmostEqual(9, e.trigger_mask)
         self.assertAlmostEqual(10, e.utc_nanoseconds)
@@ -599,7 +600,8 @@ class TestEventInfo(TestCase):
         self.assertAlmostEqual(12, e.weight_w1)
         self.assertAlmostEqual(13, e.weight_w2)
         self.assertAlmostEqual(14, e.weight_w3)
-        self.assertAlmostEqual(15, e.event_id)
+        self.assertAlmostEqual(15, e.run_id)
+        self.assertAlmostEqual(16, e.event_id)
 
     def test_from_table(self):
         e = EventInfo.from_row({
@@ -608,7 +610,6 @@ class TestEventInfo(TestCase):
             'mc_id': 3,
             'mc_t': 4,
             'overlays': 5,
-            # 'run_id': 6,
             'trigger_counter': 6,
             'trigger_mask': 7,
             'utc_nanoseconds': 8,
@@ -616,10 +617,11 @@ class TestEventInfo(TestCase):
             'weight_w1': 10,
             'weight_w2': 11,
             'weight_w3': 12,
-            'event_id': 1,
             'livetime_sec': 13,
             'n_events_gen': 14,
             'n_files_gen': 15,
+            'run_id': 16,
+            'event_id': 1,
             })
 
         self.assertAlmostEqual(0, e.det_id)
@@ -627,7 +629,7 @@ class TestEventInfo(TestCase):
         self.assertAlmostEqual(3, e.mc_id)
         self.assertAlmostEqual(4, e.mc_t)
         self.assertAlmostEqual(5, e.overlays)
-        # self.assertAlmostEqual(6, e.run_id)
+        self.assertAlmostEqual(16, e.run_id)
         self.assertAlmostEqual(6, e.trigger_counter)
         self.assertAlmostEqual(7, e.trigger_mask)
         self.assertAlmostEqual(8, e.utc_nanoseconds)
@@ -650,7 +652,6 @@ class TestEventInfo(TestCase):
             'n_events_gen': 14,
             'n_files_gen': 15,
             'overlays': 5,
-            # 'run_id': 6,
             'trigger_counter': 6,
             'trigger_mask': 7,
             'utc_nanoseconds': 8,
@@ -658,9 +659,10 @@ class TestEventInfo(TestCase):
             'weight_w1': 10.0,
             'weight_w2': 11.0,
             'weight_w3': 12.0,
+            'run_id': 16,
             'event_id': 1,
             })
-        exp = (0, 2, 13, 3, 4.0, 14, 15, 5, 6, 7, 8, 9, 10.0, 11.0, 12.0, 1,)
+        exp = (0, 2, 13, 3, 4.0, 14, 15, 5, 6, 7, 8, 9, 10.0, 11.0, 12.0, 16, 1,)
         self.assertAlmostEqual(e.serialise(), np.array(exp, e.dtype))